refactored into class component
